# api/openapi.yaml
openapi: 3.0.3
info:
  title: Avella Ecommerce API
  description: |
    RESTful API for Avella Electronics Ecommerce MVP.
    
    **Features:**
    - Product catalog with filtering and search
    - Shopping cart management
    - Order processing with stock validation
    - JWT-based authentication
    
    **Base URL:** `http://localhost:3001/api`
  version: 1.0.0
  contact:
    name: Avella Team
    email: support@avella.com
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://api.avella.com/api
    description: Production server
...
paths:
...
  /products:
    post:
      tags: [Products]
      summary: Create product
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created
...
  /cart:
    post:
      tags: [Cart]
      summary: Create new cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
                sessionId:
                  type: string
              oneOf:
                - required: [userId]
                - required: [sessionId]
      responses:
        '201':
          description: Cart created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Cart'
...
  /orders:
    get:
      tags: [Orders]
      summary: List user orders
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      summary: Create order from cart
      description: Convert cart to order, validate stock, and update inventory atomically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartToken, shippingAddress]
              properties:
                cartToken:
                  type: string
                  format: uuid
                  description: Cart token from cart creation
                userId:
                  type: string
                shippingAddress:
                  $ref: '#/components/schemas/Address'
                billingAddress:
                  $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /customers:
    get:
      tags: [Customers]
      summary: List all customers
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags: [Customers]
      summary: Create customer directly
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                firstName: { type: string }
                lastName: { type: string }
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/User' }
  /companies:
    get:
      tags: [Companies]
      summary: List companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
    post:
      tags: [Companies]
      summary: Create company
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyInput'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Company' }
  /companies/{id}:
    get:
      tags: [Companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Company' }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Companies]
      summary: Update company
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyInput'
      responses:
        '200':
          description: Company updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Companies]
      summary: Delete company
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Company deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
                companyId:
                  type: integer
                  description: Optional ID of the company this user belongs to
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User already exists
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
components:
  schemas:
...
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cart_token:
          type: string
          format: uuid
          description: Secure token for cart access
        user_id:
          type: integer
          format: int64
          nullable: true
        session_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          nullable: true
        order_number:
          type: string
          description: Human-readable order identifier
          example: ORD-2024-000001
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        subtotal:
          type: number
        tax:
          type: number
        shipping:
          type: number
        total:
          type: number
        shipping_address:
          $ref: '#/components/schemas/Address'
        billing_address:
          $ref: '#/components/schemas/Address'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Address:
      type: object
      required: [name, street, city, state, zip, country]
      properties:
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        uuid:
          type: string
          format: uuid
          description: External UUID for secure references
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [customer, admin]
        company_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT token
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        vatNumber:
          type: string
        registrationNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        vatNumber:
          type: string
        registrationNumber:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
    ProductInput:
      type: object
      required: [name, slug, price, categoryId]
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number, format: decimal }
        sku: { type: string }
        categoryId: { type: integer }
        stockQuantity: { type: integer }
        images: { type: array, items: { type: string, format: uri } }
        specifications: { type: object }
        isActive: { type: boolean }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key